requirements.txt
feedparser
requests
readability-lxml
beautifulsoup4
rich

